+++ modules/page-memory/src/integrationTest/java/org/apache/ignite/internal/pagememory/tree/ItBplusTreeReusePageMemoryTest.java
@@ -33,9 +33,9 @@ import org.apache.ignite.internal.testframework.BaseIgniteAbstractTest;
 import org.apache.ignite.lang.IgniteInternalCheckedException;
 
 /**
- * Test with reuse list.
+ * An abstract class for testing {@link BplusTree} with {@link ReuseList} using different implementations of {@link PageMemory}.
  */
-public class ItBplusTreeReuseSelfTest extends ItBplusTreeSelfTest {
+public abstract class ItBplusTreeReusePageMemoryTest extends ItBplusTreePageMemoryTest {
     /** {@inheritDoc} */
     @Override
     protected ReuseList createReuseList(int grpId,
