+++ modules/page-memory/src/integrationTest/java/org/apache/ignite/internal/pagememory/tree/inmemory/ItBplusTreeReuseVolatilePageMemoryTest.java
@@ -1,0 +1,43 @@
+package org.apache.ignite.internal.pagememory.tree.inmemory;
+
+import java.util.concurrent.TimeUnit;
+import org.apache.ignite.internal.configuration.testframework.ConfigurationExtension;
+import org.apache.ignite.internal.configuration.testframework.InjectConfiguration;
+import org.apache.ignite.internal.pagememory.PageMemory;
+import org.apache.ignite.internal.pagememory.TestPageIoRegistry;
+import org.apache.ignite.internal.pagememory.configuration.schema.UnsafeMemoryAllocatorConfigurationSchema;
+import org.apache.ignite.internal.pagememory.configuration.schema.VolatilePageMemoryDataRegionConfiguration;
+import org.apache.ignite.internal.pagememory.inmemory.VolatilePageMemory;
+import org.apache.ignite.internal.pagememory.tree.ItBplusTreeReusePageMemoryTest;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+/**
+ * Test with reuse list and {@link VolatilePageMemory}.
+ */
+@ExtendWith(ConfigurationExtension.class)
+public class ItBplusTreeReuseVolatilePageMemoryTest extends ItBplusTreeReusePageMemoryTest {
+    @InjectConfiguration(polymorphicExtensions = UnsafeMemoryAllocatorConfigurationSchema.class)
+    private VolatilePageMemoryDataRegionConfiguration dataRegionCfg;
+
+    /** {@inheritDoc} */
+    @Override
+    protected PageMemory createPageMemory() throws Exception {
+        dataRegionCfg.change(c -> c.changeInitSize(MAX_MEMORY_SIZE).changeMaxSize(MAX_MEMORY_SIZE)).get(1, TimeUnit.SECONDS);
+
+        TestPageIoRegistry ioRegistry = new TestPageIoRegistry();
+
+        ioRegistry.loadFromServiceLoader();
+
+        return new VolatilePageMemory(
+                dataRegionCfg,
+                ioRegistry,
+                PAGE_SIZE
+        );
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    protected long acquiredPages() {
+        return ((VolatilePageMemory) pageMem).acquiredPages();
+    }
+}
